lesson-10-api-layer-ai-systems/homework/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.54.3
sentence-transformers==3.3.1
faiss-cpu==1.13.2
qdrant-client==1.12.1
tiktoken==0.8.0
langfuse==2.53.7
aiosqlite==0.20.0
numpy<2.0
